backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
nibabel==5.2.1
numpy==1.26.4
scipy==1.13.1
Pillow==10.4.0
torchio==0.19.7
